fix: fixed synchronization on UserService

+++ DistLedgerServer/src/main/java/pt/tecnico/distledger/server/grpc/UserServiceImpl.java
@@ -91,9 +91,12 @@ public class UserServiceImpl extends UserServiceGrpc.UserServiceImplBase {
         if (canWrite()) {
             try {
                 String userID = request.getUserId();
-                state.createAccount(userID);
 
-                propagateState();
+                synchronized (this) {
+                    state.assertCanCreateAccount(userID);
+                    propagateState();
+                    state.createAccount(userID);
+                }
 
                 CreateAccountResponse response = CreateAccountResponse.newBuilder().build();
                 responseObserver.onNext(response);
@@ -101,7 +104,6 @@ public class UserServiceImpl extends UserServiceGrpc.UserServiceImplBase {
             } catch (AccountAlreadyExistsException e) {
                 System.err.println(ACCOUNT_ALREADY_EXISTS);
                 responseObserver.onError(Status.ALREADY_EXISTS.withDescription(ACCOUNT_ALREADY_EXISTS).asRuntimeException());
-
             } catch (ServerUnavailableException e) {
                 System.err.println(SERVER_UNAVAILABLE);
                 responseObserver.onError(Status.UNAVAILABLE.withDescription(SERVER_UNAVAILABLE).asRuntimeException());
@@ -121,9 +123,12 @@ public class UserServiceImpl extends UserServiceGrpc.UserServiceImplBase {
         if (canWrite()) {
             try {
                 String userID = request.getUserId();
-                state.deleteAccount(userID);
 
-                propagateState();
+                synchronized (this) {
+                    state.assertCanDeleteAccount(userID);
+                    propagateState();
+                    state.deleteAccount(userID);
+                }
 
                 DeleteAccountResponse response = DeleteAccountResponse.newBuilder().build();
                 responseObserver.onNext(response);
@@ -185,9 +190,12 @@ public class UserServiceImpl extends UserServiceGrpc.UserServiceImplBase {
                 String dest = request.getAccountTo();
                 Integer amount = request.getAmount();
 
-                propagateState();
+                synchronized (this) {
+                    state.assertCanTransferTo(userID, dest, amount);
+                    propagateState();
+                    state.transferTo(userID, dest, amount);
+                }
 
-                state.transferTo(userID, dest, amount);
                 TransferToResponse response = TransferToResponse.newBuilder().build();
                 responseObserver.onNext(response);
                 responseObserver.onCompleted();
